requirements.txt
flet>=0.21.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dateutil>=2.8.2 
torch>=2.0.0
transformers>=4.30.0 